webComponent created (without shadow DOM usage due to styling practices)
