+++ public/styles.css
@@ -4,7 +4,6 @@ body{
   font-family: Arial, Helvetica, sans-serif;
 }
 .nav-bar-logo{
-  /* position: -webkit-sticky; */
   height: 45px;
   width: auto;
 }
@@ -41,7 +40,6 @@ span{
 .image-outline{
     position: relative;
     width: 96vw;
-    /* animation-duration: 1.5s; */
 }
 @media (min-aspect-ratio: 9/12){
     .image-outline{
@@ -58,7 +56,6 @@ span{
         box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
         transition: 0.3s;
         border-radius: 5px;
-        /* transition: width 2s; */
     
 }
 @media (min-aspect-ratio: 9/12){
@@ -210,7 +207,7 @@ span{
     padding-bottom: 4vw;
     box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
     transition: 0.3s;
-    border-radius: 5px; /* 5px rounded corners */
+    border-radius: 5px;
   }
   @media (min-aspect-ratio: 9/12) {
       .single-image-card{
