+++ views/index.ejs
@@ -5,5 +5,5 @@
     <div id="APODwrapper" class="wrapper">
         
     </div>
-    <script src="/public/indexCtrl.js" type="text/javascript"></script>
+    <script src="/public/indexCtrl.js" type="module"></script>
     <%- include('footer') -%>
